Pag_web/requirements.txt
streamlit
python-docx
PyPDF2
openai==0.28
pandas
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
reportlab
Pillow
nltk
google-cloud-documentai